Use casts to avoid comparing differently-signed numbers.
Fixes:

    stopgp32.c:112:18: warning: comparison of integer expressions of different signedness: ‘int’ and ‘unsigned int’ [-Wsign-compare]
    stopgp32.c:125:18: warning: comparison of integer expressions of different signedness: ‘int’ and ‘unsigned int’ [-Wsign-compare]

+++ stopgp32.c
@@ -109,7 +109,7 @@ static void cache_dir_init(struct cache_dir *o)
         int size = snprintf(o->path, sizeof o->path, "%s/" PROGRAM_NAME, cache_home);
         if (size < 0)
             posix_error("$XDG_CACHE_HOME/" PROGRAM_NAME);
-        if (size >= sizeof o->path) {
+        if ((size_t) size >= sizeof o->path) {
             errno = ENAMETOOLONG;
             posix_error("$XDG_CACHE_HOME/" PROGRAM_NAME);
         }
@@ -122,7 +122,7 @@ static void cache_dir_init(struct cache_dir *o)
         int size = snprintf(o->path, sizeof o->path, "%s/.cache/" PROGRAM_NAME, home);
         if (size < 0)
             posix_error("$HOME/.cache/" PROGRAM_NAME);
-        if (size >= sizeof o->path) {
+        if ((size_t) size >= sizeof o->path) {
             errno = ENAMETOOLONG;
             posix_error("$HOME/.cache/" PROGRAM_NAME);
         }
